% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jensen-Shannon.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebysev's inequality}
\usage{
chebyshev(vec)
}
\arguments{
\item{vec}{The vector containing the values of the  data set.}
}
\value{
It outputs the lower and upper bound of Chebyshevs 75\% range and the
lower and upper value of the range of the observed (measured) distribution.
}
\description{
Function to calculate lower and upper bounds for which at least 75\% of data
points of a distribution lie within ±2SD of the mean using Chebyshev's inequality.
}
\examples{
\dontrun{
chebysevBounds = chebyshev(sampleA$LDL)
}
}
\references{
Bienayme I. Considerations a l appui de la de couverte d
laplace. Comptes Rendus de l Acade mie des Sciences 1853; 37: 309–324.

. Chebyshev P. Des valeurs moyennes. Journal de Mathematiques Pures .
Appliqee 1867; 2(12): 177–184s
}
