% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVfunctionRange.R
\name{CV}
\alias{CV}
\title{Calculate the CV of a set of values.}
\usage{
CV(Vec, roundDigits = 2)
}
\arguments{
\item{Vec}{The vector of values for which to calculate the CV.}

\item{roundDigits}{Number of digits to round the result to (default=2).}
}
\value{
Returns the value of the CV.
}
\description{
This function calculates the coefficient of variation (CV) of the values supplied.
}
\examples{
CV = CV(sampleA$LDL)
}
