% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{DensityPlotOfVar}
\alias{DensityPlotOfVar}
\title{Plot variance density}
\usage{
DensityPlotOfVar(Vector, title = "")
}
\arguments{
\item{Vector}{A vector or data frame columns with the values whose}

\item{title}{Title of the plot (default="").
distribution density is to be plotted.}
}
\value{
The function returns no values but plots the density plot.
}
\description{
Plot density of the Variance along with the median, 2.5 and 97.5
percentile as vertical lines
}
\examples{
\dontrun{
LDLbootstrVar=as.data.frame(LDLbootVrnc(sampleA$CHOL,sampleA$HDL,sampleA$TG))
DensityPlotOfVar(LDLbootstrVar$dataTable.CV)
}
}
