% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fErrorCorCov.R
\name{ErrorCorCov}
\alias{ErrorCorCov}
\title{Calculate correlation and covariance of residuals}
\usage{
ErrorCorCov(dataset1, dataset2, plot = F)
}
\arguments{
\item{dataset1}{A vector containing the values of the first group.}

\item{dataset2}{A vector containing the values of the second group.}

\item{plot}{(default=TRUE). Plot the errors of the first group versus the second one.}
}
\value{
Returns a list with the residuals of the two groups (distr1Error,
distr2Error), the correlation coefficient between the two groups (Correlation)
and the covariance of the two groups (Covariance).
}
\description{
This function calculates the residuals
(the squared difference of each value from the mean) of two groups. Then
it calculates the correlation and covariance between the residuals of the two
groups and plots them.
}
\examples{
\dontrun{
ErrorOFCorCov=ErrorCorCov(sampleA$HDL[1:20],sampleA$CHOL[1:20],plot = FALSE)
 }
}
