% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equation_functions.R
\name{LDL_eq}
\alias{LDL_eq}
\title{Calculates and returns the LDL Value for any of the 12 different equations}
\usage{
LDL_eq(TC, HDL, TG, EqMethod)
}
\arguments{
\item{TC}{The TC (Total Cholesterol) value.}

\item{HDL}{The HDL (High-density lipoprotein- cholesterol) value.}

\item{TG}{The TG (Triglyceride) value.}

\item{EqMethod}{The type of equation to be used to calculate the LDL value.The type of equation to be used to calculate the LDL value.
EqMethod could be:("Friedewald","Ahmadi","Chen","Anandaraja","NewFormula","deCordova","Vujovic","Hattori","Puavillai","Hatta","Martin180","Martin360","Martin2000","DeLong" or "Rao").}
}
\value{
The calculated LDL value, according to the equation of choice or a printed error message and 404,
if the equation type does not exist.
}
\description{
This function calculates and returns the LDL value computed from any of the 12 named equations.
}
\examples{
LDL_eq(170.5,35.12,230,"Martin360")
}
