% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9ML_models_functions.R
\name{LDL_ML_predict}
\alias{LDL_ML_predict}
\title{Predict LDL value(s)}
\usage{
LDL_ML_predict(model, data)
}
\arguments{
\item{model}{The model with which the predictions will be made.}

\item{data}{The data with which the predictions will be made, can either be a single set of (CHOL,HDL,TG) values
or a data table of sets of said values.}
}
\value{
The predicted LDL value(s).
}
\description{
This function predicts and returns  predictions, based on the model previously trained.
}
\examples{
modelPrediction = LDL_ML_predict(model,data.table::data.table(CHOL=170.5,HDL=35.12,TG=175))

}
