% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All9Models.R
\name{LDL_ML_predict_all_models}
\alias{LDL_ML_predict_all_models}
\title{Predict LDL values}
\usage{
LDL_ML_predict_all_models(models, validationData)
}
\arguments{
\item{models}{The models with which the predictions will be made.}

\item{validationData}{The data with which the predictions will be made, can either be a single set of
(CHOL,HDL,TG) values or a data table of sets of said values.}
}
\value{
The predicted LDL values.
}
\description{
This function predicts and returns your predictions, based on all the model you have previously trained.
}
\examples{
allModelsPredictions = LDL_ML_predict_all_models(allModels,data.table(CHOL=170.5,HDL=35.12,TG=175))
allModelsPredictions = LDL_ML_predict_all_models(allModels,Data)
}
