% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9ML_models_functions.R
\name{LDL_predict}
\alias{LDL_predict}
\title{Predict LDL value(s)}
\usage{
LDL_predict(model, data)
}
\arguments{
\item{model}{The model with which the predictions will be made.}

\item{data}{The data with which the predictions will be made, can either be a single set of (CHOL,HDL,TG) values
or a data table of sets of said values.}
}
\value{
The predicted LDL value(s).
}
\description{
This function predicts and returns your predictions, based on the model you previously trained.
}
\examples{
modelPrediction=LDL_predict(model,c(data.table(CHOL=170,HDL=35,TG=174)))
modelPrediction=LDL_predict(model,Data)
modelPrediction=LDL_predict(stackModel,Data)
}
