% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equation_functions.R
\name{LDLallEq}
\alias{LDLallEq}
\title{Calculates and returns the LDL values using all available equations}
\usage{
LDLallEq(TC, HDL, TG)
}
\arguments{
\item{TC}{The TC (Total Cholesterol) value.}

\item{HDL}{The HDL (High-density lipoprotein-cholesterol) value.}

\item{TG}{The TG (Triglyceride) value.}
}
\value{
The calculated LDL values, according to all the equations.
}
\description{
This function calculates and returns the LDL values computed with all of the 12 named equations.
}
\examples{
LDLallEq(170,35,174)
}
