% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotCorrWithRegrLine}
\alias{PlotCorrWithRegrLine}
\title{Scatterplot of pairs of parameters}
\usage{
PlotCorrWithRegrLine(df, xParam, yParam)
}
\arguments{
\item{df}{The data frame with the parameterrs to be plotted}

\item{xParam}{The parameter (column name) to be plotted in the abscissa (x axis).}

\item{yParam}{The prameter (column name) to be plotted in the ordinate (y axis).}
}
\value{
The function returns a ggplot2 object.
}
\description{
Scatterplot of pairs of parameters with the corresponding regression line.
}
\examples{
\dontrun{
PlotCorrWithRegrLine(sampleA,"CHOL", "HDL")
}
}
