% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotDiscrHist}
\alias{PlotDiscrHist}
\title{Plot discrete histogram}
\usage{
PlotDiscrHist(DF, param, title = "")
}
\arguments{
\item{DF}{A data frame containing columns whose discrete histogram is to be plotted.}

\item{param}{The column name of the columns for which to plot the discrete histogram.}

\item{title}{Title of the plot (default="").}
}
\value{
It returns a ggplot object.
}
\description{
Plot discrete histogram (barplot) of the data frame column named in the param
argument. It also plots the mean as a vertical continuous line, the mean plus/minus 2
standard deviations as veritcal dotted lines and overlays a density plot of the
normal distribution with mean and standard deviation corresponding to those of the data.
}
\examples{
\dontrun{
PlotDiscrHist(sampleA,"LDL")
}
}
