% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{ReportMultiPlot}
\alias{ReportMultiPlot}
\title{Creates a multi-plot diagram reporting very useful information about the model}
\usage{
ReportMultiPlot(inputModelPredict, validationDataColumn, name = "")
}
\arguments{
\item{inputModelPredict}{The model's predictions.}

\item{validationDataColumn}{The validation data (the actual values that had to be predicted).}

\item{name}{A string as the diagram subtitle}
}
\value{
A diagram with 5 plots.
}
\description{
This function creates and returns a full diagram of 4 different plots, relying information about a prediction model.
}
