% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleData.R
\docType{data}
\name{SampleData}
\alias{SampleData}
\title{Anonymous real data}
\format{
A \code{tibble} with 5 elements, which are:
\describe{
\item{AGE}{The Age of cases}
\item{CHOL}{The Cholesterol of cases}
\item{TG}{The Triglyceride of cases}
\item{HDL}{The High Density Lipoprotein of cases}
\item{LDLd}{The direct Low Density Lipoprotein of cases}
}
}
\usage{
SampleData
}
\description{
A dataset \code{tibble} for the test of your model, containing values of 2000 cases. It is used as data for the examples.
}
\keyword{datasets}
